fix(title): fix copying on WSL 2 by bumping clipboardy version
In [this PR](https://github.com/sindresorhus/clipboardy/pull/59),
support for WSL was added in `clipboardy` v2.3.0. Since WSL
support for `title` would be nice, `clipboardy`'s version was increased
from 1.2.2 to 3.0.0.

Because `clipboardy` is now an ES module, the method of importing it had
to change to use the `import` function rather than `require`.