build(pnpm)!: finish transition to `pnpm`
What motivated the change is that it wasn't clear from the `readme.md`
nor the CI job that `pnpm` is the main package manager for the project.

In addition to updating the CI job & docs, the `packageManager` field of
the `package.json` was updated to help signify the intended package
manager used to develop the project.

BREAKING CHANGE: Dropped support for Node.js versions lower than 18.
Node.js 16's security updates ended on 2023-09-11.

Refs: https://github.com/nodejs/Release#end-of-life-releases